Move NumberFingerprint to its own file
