+++ frontend/src/metabase/components/MetadataInfo/FieldFingerprintInfo/FieldFingerprintInfo.tsx
@@ -1,5 +1,3 @@
-import { t } from "ttag";
-
 import type { DatasetColumn } from "metabase-types/api";
 import * as Lib from "metabase-lib";
 import {
@@ -9,10 +7,10 @@ import {
   isNumber,
 } from "metabase-lib/types/utils/isa";
 import type Field from "metabase-lib/metadata/Field";
-import { Table } from "../MetadataInfo.styled";
 
 import CategoryFingerprint from "./CategoryFingerprint";
 import { DateTimeFingerprint } from "./DateTimeFingerprint";
+import { NumberFingerprint } from "./NumberFingerprint";
 
 type FieldFingerprintInfoProps = {
   className?: string;
@@ -89,58 +87,6 @@ function FieldFingerprintInfo({
   }
 
   return null;
-}
-
-/**
- * @param num - a number value from the type/Number fingerprint; might not be a number
- * @returns - a tuple, [isFormattedNumber, formattedNumber]
- */
-function roundNumber(num: number | null | undefined): [boolean, string] {
-  if (num == null) {
-    return [false, ""];
-  }
-
-  return [true, Number.isInteger(num) ? num.toString() : num.toFixed(2)];
-}
-
-type NumberFingerprintProps = {
-  className?: string;
-  fingerprintTypeInfo?: Lib.NumberFingerprintDisplayInfo | null;
-};
-
-function NumberFingerprint({
-  className,
-  fingerprintTypeInfo,
-}: NumberFingerprintProps) {
-  if (!fingerprintTypeInfo) {
-    return null;
-  }
-
-  const { avg, min, max } = fingerprintTypeInfo;
-  const [isAvgNumber, formattedAvg] = roundNumber(avg);
-  const [isMinNumber, formattedMin] = roundNumber(min);
-  const [isMaxNumber, formattedMax] = roundNumber(max);
-
-  const someNumberIsDefined = isAvgNumber || isMinNumber || isMaxNumber;
-
-  return someNumberIsDefined ? (
-    <Table className={className}>
-      <thead>
-        <tr>
-          {isAvgNumber && <th>{t`Average`}</th>}
-          {isMinNumber && <th>{t`Min`}</th>}
-          {isMaxNumber && <th>{t`Max`}</th>}
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          {isAvgNumber && <td>{formattedAvg}</td>}
-          {isMinNumber && <td>{formattedMin}</td>}
-          {isMaxNumber && <td>{formattedMax}</td>}
-        </tr>
-      </tbody>
-    </Table>
-  ) : null;
 }
 
 // eslint-disable-next-line import/no-default-export
